backend/requirements.txt
scrapy==2.11.1
scrapy-playwright==0.0.33
itemadapter==0.9.0
pymongo==4.6.1
python-dotenv==1.0.1
price-parser==0.3.4
textblob==0.18.0
fake-useragent==1.4.0
fastapi==0.115.6
uvicorn==0.34.0
pydantic==2.10.4
